removed unnecessary holder and renamed auth components
